@@ -1,10 +1,12 @@
 package eu.long1.spacetablayoutdemo;
 
+
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.fragment.app.Fragment;
 
 public class FragmentA extends Fragment {
 
